Rename Core.Core.* to Core.*.
